+++ examples/Example4-BigExampleFromPaper.txt
