+++ examples/Example4.1-BigExampleFromPaperFewerLeaves.txt
