+++ examples/Example4.2-BigExampleFromPaperReduced.txt
