+++ src/phyloroot/network_rootability.py
@@ -1,4 +1,4 @@
-from Rooting import *
+from rooting import *
 import networkx as nx
 import os
 import sys
